% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-solver.R
\name{helper_solver}
\alias{helper_solver}
\title{Helper solver}
\usage{
helper_solver(inputs, ..., solver_chat)
}
\arguments{
\item{inputs}{List of input objects from the helperbench dataset. Each input
should have the refactoring prompt.}

\item{...}{Additional arguments (currently unused).}

\item{solver_chat}{An ellmer Chat object to use for solving the prompts.}
}
\value{
A list with the following components:
\describe{
\item{result}{Character vector of model responses, one for each input.}
\item{solver_chat}{List of Chat objects used to generate each response.}
\item{solver_metadata}{List containing the solver working directory for each
input, which will be used and cleaned up by the scorer.}
}
}
\description{
Pass this function to \code{Task$new()} as the solver to process inputs from
\link{helper_dataset} with a specified language model.
}
\details{
The solver creates isolated working directories for each input, initializes
side::kick() clients with appropriate tools, and prompts models to perform
the refactoring task. Results include both the model's response and the
modified R/ directory.
}
